# specs/001-qr-attendance-payroll/contracts/api.yaml
openapi: 3.1.0
info:
  title: AttendanceApp API
  version: 1.0.0
  description: |
    REST API for the QR-based employee attendance and payroll system.
    All endpoints (except /auth/login) require a valid Bearer JWT.
    All communication MUST use HTTPS/TLS 1.2+.
servers:
  - url: https://attendanceapp-api.onrender.com/api
    description: Production (Render.com)
  - url: http://localhost:3000/api
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:    { type: string }
        message: { type: string }
        details: { type: object }
    Employee:
      type: object
      properties:
        id:          { type: string }
        name:        { type: string }
        email:       { type: string, format: email }
        role:        { type: string, enum: [employee, admin] }
        status:      { type: string, enum: [active, inactive] }
        workplaceId: { type: string }
        createdAt:   { type: string, format: date-time }
    AttendanceRecord:
      type: object
      properties:
        id:                   { type: string }
        employeeId:           { type: string }
        date:                 { type: string, format: date, example: "2026-02-13" }
        checkInAt:            { type: string, format: date-time }
        checkOutAt:           { type: string, format: date-time, nullable: true }
        status:               { type: string, enum: [open, closed, incomplete] }
        regularHours:         { type: number, nullable: true }
        overtimeHoursMorning: { type: number, nullable: true }
        overtimeHoursEvening: { type: number, nullable: true }
        isManuallyAdjusted:   { type: boolean }
        adjustmentNote:       { type: string, nullable: true }
...
security:
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [auth]
      summary: Authenticate and receive tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:    { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:  { type: string }
                  employee:     { $ref: '#/components/schemas/Employee' }
          headers:
            Set-Cookie:
              schema:
                type: string
                description: HttpOnly refresh token cookie (web); also returned in body for mobile
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/refresh:
    post:
      tags: [auth]
      summary: Issue new access token using refresh token
      security: []
      description: |
        Mobile: send refreshToken in request body.
        Web: send via HttpOnly cookie (automatically included by browser).
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string, description: "Mobile only — omit for web (uses cookie)" }
      responses:
        '200':
          description: New access token issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken: { type: string }
        '401':
          description: Refresh token invalid or expired
...
  /attendance/checkin:
    post:
      tags: [attendance]
      summary: Record employee check-in via QR scan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [qrToken, deviceTimestamp, location]
              properties:
                qrToken:
                  type: string
                  description: UUID from the workplace QR code
                deviceTimestamp:
                  type: string
                  format: date-time
                  description: Device time at moment of scan (for audit; server time is authoritative)
                location:
                  type: object
                  required: [lat, lng]
                  properties:
                    lat: { type: number }
                    lng: { type: number }
      responses:
        '201':
          description: Check-in recorded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  record:         { $ref: '#/components/schemas/AttendanceRecord' }
                  serverCheckInAt: { type: string, format: date-time }
        '400':
          description: Invalid QR token or already checked in today
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: Employee not assigned to this workplace
  /attendance/checkout:
    post:
      tags: [attendance]
      summary: Record employee check-out via QR scan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [qrToken, deviceTimestamp, location]
              properties:
                qrToken:         { type: string }
                deviceTimestamp: { type: string, format: date-time }
                location:
                  type: object
                  properties:
                    lat: { type: number }
                    lng: { type: number }
      responses:
        '200':
          description: Check-out recorded; returns closed record with pay summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  record:            { $ref: '#/components/schemas/AttendanceRecord' }
                  serverCheckOutAt:  { type: string, format: date-time }
                  summary:
                    type: object
                    properties:
                      totalHours:           { type: number }
                      regularHours:         { type: number }
                      overtimeHoursMorning: { type: number }
                      overtimeHoursEvening: { type: number }
        '400':
          description: No open check-in found for today or invalid QR
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /attendance/sync:
    post:
      tags: [attendance]
      summary: Bulk sync offline attendance events
      description: |
        Used when the mobile app reconnects after offline period.
        Each event is processed in order; duplicate events return 200 with a conflict flag.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [events]
              properties:
                events:
                  type: array
                  items:
                    type: object
                    required: [type, qrToken, deviceTimestamp, location, localId]
                    properties:
                      localId:         { type: string, description: "Client-generated UUID for dedup" }
                      type:            { type: string, enum: [checkin, checkout] }
                      qrToken:         { type: string }
                      deviceTimestamp: { type: string, format: date-time }
                      location:
                        type: object
                        properties:
                          lat: { type: number }
                          lng: { type: number }
      responses:
        '200':
          description: Sync results per event
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        localId: { type: string }
                        status:  { type: string, enum: [synced, conflict, error] }
                        record:  { $ref: '#/components/schemas/AttendanceRecord' }
                        message: { type: string }

  # ─── ADMIN — EMPLOYEES ───────────────────────────────────────────────────────
...